requirements.txt
fastapi
uvicorn
langchain
python-dotenv
langchain_openai